Move shroud traits to Mods.Common.
