+++ OpenRA.Mods.Common/OpenRA.Mods.Common.csproj
@@ -681,6 +681,8 @@
     <Compile Include="Traits\Modifiers\HiddenUnderShroud.cs" />
     <Compile Include="Lint\CheckDefaultVisibility.cs" />
     <Compile Include="Traits\Modifiers\AlwaysVisible.cs" />
+    <Compile Include="Traits\CreatesShroud.cs" />
+    <Compile Include="Traits\RevealsShroud.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <PropertyGroup>
